feat: add mcp-server to preview workflows
This adds a new tool that allows the LLM to render a workflow for the user. When the LLM determines that everything is ready, it will provide the user with a workflow display to help them understand the process.

This implementation is somewhat complex because SWFEditor always depends on a browser, and Java classes are not available in Maven Central. I've taken an approach that renders using static HTML and extracts it using a headless browser. While not optimal for performance, this solution works until we can make changes to the Stunner editor.

Related repositories:
- https://github.com/apache/incubator-kie-tools/tree/main/packages/
- Editor: https://github.com/apache/incubator-kie-tools/tree/ab3030b602d74d2f30227cba5c2c88aaccf2d2a8/packages/serverless-workflow-standalone-editor

Signed-off-by: Eloy Coto <eloy.coto@acalustra.com>